principalEmzima/pythonAutomatizacion/pythonCucumber/requirements.txt
pytest==8.3.3
selenium==4.25.0
webdriver-manager==4.0.1
pytest-html==4.1.1
allure-pytest==2.13.5
python==3.12.6
pytest-bdd==7.3.0